backend/requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
sqlalchemy>=2.0.0